+++ CHANGES.rst
@@ -73,8 +73,8 @@ Changelog
   PLONE_FIXTURE.
   [timo]
 
-- Fix ownership of comments. [toutpt]
-
+- Fix ownership of comments.
+  [toutpt]
 
 2.2.10 (2013-09-24)
 -------------------